Allow the installer to be run without elevated privileges, and enable single-user install mode
Now, when a user runs the installer without elevated privileges it will perform a per-user
installation by default, to a default path under the user's $USERPROFILE (over which the
user has full control).  A site-wide install under Program Files is selectable (and is the
default) if the installer is run with elevated privileges.