Enhance subcyg command: Use some clever hacks to terminate the wrapped command when hitting Ctrl-C
Previously, a Ctrl-C would just interrupt the 'tail' command that was being
used to follow the output from the subcyg shell.  But now if 'tail' is
interrupted we follow through by running another shell in the subcyg but
with a kill command to kill the process group of the original wrapped
command.