Improvements to the subcyg script:
It now tries to better handle interrupting a "subcyg" command with
Ctrl-C.  Previously an interrupt/Ctrl-C would just cancel the `tail`
command that streams the output from the subcyg command, while the
underlying command would continue to run.  This now tries to also kill
the underlying command.  It still doesn't seem to always work perfectly
but it's better than before.

Commands passed to subcyg are also now automatically run in a subshell
by wrapping them in (), so it's not necessary to do this explicitly
anymore.