Add additional build dependencies to the runtime environment
This is also needed to prevent packages already provided by the system
from being rebuilt when installing optional packages.

In the future it would be nice (assuming we don't already have one) to have
a way to forcibly disable installation of dependencies as SPKGs, even if
configure *thinks* they are unavailable.