Support building from a local Cygwin mirror.
In a follow-up, the plan is to commit the (partial) mirror I used for
this build to a repository, so that builds can be repeated more easily
in the future.

The problem is that Cygwin's official mirrors have rolling versions, and
only maintain the last 2 or 3 versions of each package.  It's also
difficult to direct the installer to use older versions of packages.
This leads sometimes to unpredictable bugs where building a new version
of Sage for Windows uses a newer version of Cygwin that has new bugs.

Also note that the latest build is based on an older version of Cygwin
in order to work around a bug in the most current version.